% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{as_symbolic}
\alias{as_symbolic}
\title{Convert a Function to Symbolic Class}
\usage{
as_symbolic(.f)
}
\arguments{
\item{.f}{A function to be converted into a symbolic function.}
}
\value{
A function with an added "symbolic" class, allowing it to be used in symbolic transformations.
}
\description{
This function takes a regular function and converts it into a symbolic function by assigning it a class of "symbolic".
It is intended to prepare functions for symbolic transformations using the \code{symbolic} package.
}
\examples{
.f <- function(x) x^2
symbolic_f <- as_symbolic(.f)
class(symbolic_f)  # Returns "symbolic"

}
