% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic-package.R
\docType{package}
\name{symbolic-package}
\alias{symbolic}
\alias{symbolic-package}
\title{symbolic: Computer Algebra System for Symbolic Functional Manipulation}
\description{
Provides symbolic algebra capabilities for R. Includes functions for symbolic manipulation, simplification, differentiation, integration, equation solving, and more. Ideal for users working with algebraic expressions and symbolic computation in mathematics and scientific applications.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
